+++ app/layout.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from "next";
 import { Inter } from "next/font/google";
 import "./globals.css";
 import Head from "next/head";
+import { TooltipProvider } from "@/components/ui/tooltip";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -23,7 +24,9 @@ export default function RootLayout({
           content="width=device-width, initial-scale=1, maximum-scale=1"
         ></meta>
       </Head>
-      <body className={inter.className}>{children}</body>
+      <TooltipProvider>
+        <body className={inter.className}>{children}</body>
+      </TooltipProvider>
     </html>
   );
 }
