+++ components/helper/constants.ts
@@ -1,0 +1,58 @@
+export const FETCH_URL = [
+  `https://sewacrowd.sonnylazuardi.workers.dev/`,
+  `https://opensheet.elk.sh/1ibbywsC1tQ_sLPX8bUAjC-vrTrUqZgZA46W_sxWw4Ss/Approved+Datasheets+(Monitor)`,
+];
+
+export const TASKS = [
+  "Wordnet",
+  "Machine Translation",
+  "Readibility Assessment",
+  "Automatic Speech Recognition",
+  "Question Answering",
+  "Image-to-Text Generation",
+  "Language Modeling",
+  "Text-to-Image Generation",
+  "Visual Storytelling",
+  "Image Captioning",
+  "Speech Emotion Recognition",
+  "Aspect Based Sentiment Analysis",
+  "Text Retrieval",
+  "Text-To-Speech Synthesis",
+  "Optical Character Recognition",
+  "Commonsense Reasoning",
+  "Dialogue System",
+  "Abstractive Summarization",
+  "Sentiment Analysis",
+  "Speech-to-text translation",
+  "Emotion Classification",
+  "Hate Speech Detection",
+  "Error Spelling Correction",
+  "Text Classification",
+  "Code-switching Identification",
+  "Summarization",
+  "Relation Extraction",
+  "Named Entiy Recognition",
+  "Spoken Language Identification",
+  "Language Identification",
+  "Video-to-Text Retrieval",
+  "Video Captioning",
+  "Chatbot",
+  "Paraphrasing",
+  "Constituency Parsing",
+  "Short Answer Grading",
+  "Coreference Resolution",
+  "POS Tagging",
+  "Token-level language identification",
+  "Abstract Meaning Representation",
+  "Legal Classification",
+  "Next Tweet Prediction",
+  "Tweet Ordering",
+  "Hoax Detection",
+  "Clickbait Detection",
+  "Topic Classification",
+  "Stance Detection",
+  "Semantic Role Labeling",
+  "Intent Classification",
+  "Slot Filling",
+  "Abusive Language Detection",
+];
